+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Added
     (e.g., STL or MLT). Example usage:
 
     .. code:: python
+
         # Create task for each relation.
         tasks = create_task(
             task_names = TASK_NAMES,