+++ CHANGELOG.rst
@@ -1,6 +1,9 @@
 Unreleased_
 -----------
 
+0.8.1_ - 2020-04-13
+-------------------
+
 Added
 ^^^^^
 * `@senwu`_: Add `mode` argument in create_task to support `STL` and `MTL`.
@@ -1095,7 +1098,8 @@ Added
 ^^^^^
 * `@lukehsiao`_: Deploy Fonduer to PyPi using Travis-CI
 
-.. _Unreleased: https://github.com/hazyresearch/fonduer/compare/v0.8.0...master
+.. _Unreleased: https://github.com/hazyresearch/fonduer/compare/v0.8.1...master
+.. _0.8.1: https://github.com/hazyresearch/fonduer/compare/v0.8.0...v0.8.1
 .. _0.8.0: https://github.com/hazyresearch/fonduer/compare/v0.7.1...v0.8.0
 .. _0.7.1: https://github.com/hazyresearch/fonduer/compare/v0.7.0...v0.7.1
 .. _0.7.0: https://github.com/hazyresearch/fonduer/compare/v0.6.2...v0.7.0